requirements.txt
flask
flask-bootstrap
flask-wtf
Pillow
opencv-python
pymupdf
requests